Making the example clearer
